Move targeting logic into a library and test it
The binary is still there, but amounts to calling into the library.

I'm still not testing the database write, but this is a start.
